Update alogProcessor.mlx
Updated with the objective and description of the script.

Defined the desired variables.

Automated the retrieval process for fetching variables for a given mission.

Automated combining of GPS time into a single matrix

Automated combining of NAV LAT/LONG into a single matrix.

Working on combining GPS time with NAV LAT/LONG data.